+++ src/main/java/io/paradaux/friendlybot/commands/staff/technician/DmCommand.java
@@ -26,7 +26,7 @@ package io.paradaux.friendlybot.commands.staff.technician;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import io.paradaux.friendlybot.managers.PermissionManager;
 import io.paradaux.friendlybot.utils.models.configuration.ConfigurationEntry;
-import io.paradaux.friendlybot.utils.models.objects.PrivilegedCommand;
+import io.paradaux.friendlybot.utils.models.types.PrivilegedCommand;
 import org.slf4j.Logger;
 
 public class DmCommand extends PrivilegedCommand {
