+++ src/main/java/io/paradaux/friendlybot/commands/utility/UserInfoCommand.java
@@ -28,7 +28,7 @@ import io.paradaux.friendlybot.managers.PermissionManager;
 import io.paradaux.friendlybot.utils.embeds.command.UserInfoEmbed;
 import io.paradaux.friendlybot.utils.StringUtils;
 import io.paradaux.friendlybot.utils.models.configuration.ConfigurationEntry;
-import io.paradaux.friendlybot.utils.models.objects.PrivilegedCommand;
+import io.paradaux.friendlybot.utils.models.types.PrivilegedCommand;
 import net.dv8tion.jda.api.entities.Member;
 import org.slf4j.Logger;
